Containers/requirements.txt
feedparser==6.0.10
mysql-connector-python==8.0.33
python-dotenv==1.0.0
schedule==1.1.0
requests==2.31.0
beautifulsoup4==4.12.2
